+++ .env_EXAMPLE
@@ -1,0 +1,1 @@
+DB_PASSWORD=
