+++ internal/delivery/http/v1/auth.go
@@ -1,4 +1,4 @@
-package handler
+package v1
 
 import (
 	"net/http"
