+++ internal/delivery/http/v1/habit_handler.go
@@ -1,4 +1,4 @@
-package handler
+package v1
 
 import (
 	"fmt"
