+++ internal/delivery/http/v1/middleware.go
@@ -1,4 +1,4 @@
-package handler
+package v1
 
 import (
 	"errors"
