+++ internal/delivery/http/v1/response.go
@@ -1,4 +1,4 @@
-package handler
+package v1
 
 import (
 	"github.com/gin-gonic/gin"
